Update README.md
The DER index, as opposed to the Esteban and Ray, (1994) index, takes a continuous variable (typically income) as the polarization source, rather than a categorical variable. The index, instead of dividing individuals into groups defined by some categorical variable, takes into account the presence of clusters of the continuous variable. From the distribution point of view, the presence of clusters implies a multi peaked shape. Because of the above, the index tends to increase when the peaks of the distribution are more distant, and also tends to increase if the peak densities become larger.

The intuition behind the index, is that distant peaks of the distribution create "alienation" betweent individuals around those peaks, and larger peak densities imply that a larger amount of individuals are alienated from those around another peak.

An example of this index computation can be found in (Eggers and López-Marmolejo, 2020), for the Central-American region: https://publications.iadb.org/es/polarizacion-instituciones-y-conflicto-una-aplicacion-mexico-el-istmo-centroamericano-y-republica